Use platform_set_tls_base() throughout, remove duplication
Introduce platform_set_tls_base() throughout hvt/muen/virtio/spt
bindings and remove duplication with code in crt_init.h.
