+++ bindings/GNUmakefile
@@ -29,7 +29,7 @@ all_TARGETS :=
 $(V).SILENT:
 
 common_SRCS := abort.c cpu_$(CONFIG_ARCH).c cpu_vectors_$(CONFIG_ARCH).S \
-    crt.c printf.c intr.c lib.c mem.c exit.c log.c cmdline.c
+    crt.c printf.c intr.c lib.c mem.c exit.c log.c cmdline.c tls.c
 
 common_hvt_SRCS := hvt/start.c hvt/platform.c hvt/platform_intr.c hvt/time.c
 
@@ -37,7 +37,7 @@ hvt_SRCS := $(common_SRCS) $(common_hvt_SRCS) \
     hvt/platform_lifecycle.c hvt/yield.c hvt/tscclock.c hvt/console.c \
     hvt/net.c hvt/block.c
 
-spt_SRCS := abort.c crt.c printf.c lib.c mem.c exit.c log.c cmdline.c \
+spt_SRCS := abort.c crt.c printf.c lib.c mem.c exit.c log.c cmdline.c tls.c \
     spt/bindings.c spt/block.c spt/net.c spt/platform.c spt/start.c \
     spt/sys_linux_$(CONFIG_ARCH).c
