+++ bindings/crt_init.h
@@ -29,48 +29,9 @@ extern uintptr_t SSP_GUARD;
 #endif
 
 /*
- * These functions are responsible for performing early initialisation required
- * by the "C runtime". They must be inlined as early as possible once in C
- * code, before calling any other functions. The calling function must not
- * return.
- *
- * As crt_init_tls() currently requires that the bindings are running in CPU
- * Ring 0 or equivalent, this is split into a separate function so that
- * bindings such as "spt" can omit the call.
+ * This function must be inlined as early as possible once in C code, before
+ * calling any other functions. The calling function must not return.
  */
-
-__attribute__((always_inline)) static inline void crt_init_tls(void)
-{
-    /*
-     * Explicitly disable any accidental use of TLS by zeroing the relevant
-     * CPU registers, thus causing TLS access to point to the zero page.
-     *
-     * On x86_64 these MSRs "should" technically be zero on hvt VM entry, but
-     * this is not actually specified anywhere. For virtio the bootloader
-     * and/or platform may have likewise clobbered the values.
-     *
-     * On aarch64 the defaults seen in practice on KVM appear to be garbage
-     * values. Note that we are running at EL1, but based on practical
-     * investigation TPIDR_EL0 is the register that the compiler would use to
-     * point to TLS.
-     */
-#if defined(__x86_64__)
-     __asm__ __volatile("wrmsr" ::
-         "c" (0xc0000100), /* IA32_FS_BASE */
-         "a" (0x0),
-         "d" (0x0)
-     );
-     __asm__ __volatile("wrmsr" ::
-         "c" (0xc0000101), /* IA32_GS_BASE */
-         "a" (0x0),
-         "d" (0x0)
-     );
-#elif defined(__aarch64__)
-     __asm__ __volatile("msr tpidr_el0, xzr");
-#else
-#error Unsupported architecture
-#endif
-}
 
 __attribute__((always_inline)) static inline void crt_init_ssp(void)
 {
@@ -79,4 +40,14 @@ __attribute__((always_inline)) static inline void crt_init_ssp(void)
      */
     SSP_GUARD = READ_CPU_TICKS() + (READ_CPU_TICKS() << 32UL);
     SSP_GUARD &= ~(uintptr_t)0xff00;
+}
+
+/*
+ * Explicitly disable any accidental use of TLS by setting the arch-specific
+ * TLS base to zero, thus causing TLS access to point to the zero page.
+ */
+
+__attribute__((always_inline)) static inline void crt_init_tls(void)
+{
+    (void)platform_set_tls_base(0);
 }
