+++ bindings/spt/bindings.c
@@ -47,8 +47,4 @@ solo5_time_t solo5_clock_wall(void)
     return (ts.tv_sec * NSEC_PER_SEC) + ts.tv_nsec;
 }
 
-solo5_result_t solo5_set_tls_base(uintptr_t base)
-{
-    int rc = platform_set_tls_base(base);
-    return (rc == 0) ? SOLO5_R_OK : SOLO5_R_EINVAL;
-}
+/* solo5_set_tls_base is in tls.c */
