+++ bindings/spt/start.c
@@ -23,8 +23,8 @@
 
 void _start(void *arg)
 {
-    platform_init_tls();
     crt_init_ssp();
+    crt_init_tls();
 
     static struct solo5_start_info si;
