+++ bindings/virtio/start.c
@@ -26,8 +26,8 @@ static void _start2(void *arg) __attribute__((noreturn));
 
 void _start(void *arg)
 {
-    crt_init_tls();
     crt_init_ssp();
+    crt_init_tls();
 
     volatile int gdb = 1;
